Mecanum Inference is now in alpha!
Like many Haskell programmers, none of mecanuminference.h has been
tested. FIXME: test it. However, it should *in theory* work --
everything type-checks and sanity checks from what I can tell.
Everything is written in a modular functional style of C as well, so it
should [in theory] be easy to fix any issues with it. Basically,
mecanuminference.h "simulates" a mecanum drive so it's possible to
track approximate position, velocity and heading of a mecanum robot
without a gyro and an accelerometer. The same code could also
theoretically be used to create an application to train in new drivers
without requiring a mecanum drive.