Fixed issues with floats, fixed the math for "simplemecanum"
Changed every occurence of "abs" in math to "fabs" [more correct],
corrected the "angle" method of vectormath.h so it always returns
positive angle measures, verified and fixed everything in
"simplemecanum.h" -- full output is now given if demanded. Added
mecanum.cpp, which was used to stress-test mecanumdrive.h. Now that the
math is fixed, the code can be consolidated.