Inference header revised
Recognized that I was being stupid on one dimension by approximating an
integral that doesn't need to be approximated, so I added a helper
function. Then recognized that velocity is currently independent of
rotation -- clever kludges to fix it will be coming your way when I
have more free time.